refactor: 참조 오류 해결
