+++ src/pages/AssetManagement/pages/RegularAsset/components/DeleteButton/index.ts
@@ -1,1 +1,3 @@
+import DeleteButton from "@pages/AssetManagement/pages/RegularAsset/components/DeleteButton/DeleteButton.tsx";
+
 export default DeleteButton;
