+++ src/pages/AssetManagement/pages/SavingGoal/components/GoalSetting/index.ts
@@ -1,1 +1,3 @@
+import GoalSetting from "@pages/AssetManagement/pages/SavingGoal/components/GoalSetting/GoalSetting.tsx";
+
 export default GoalSetting;
